requirements.txt
opentelemetry-sdk>=1.40.0